requirements.txt
Flask==3.0.0
python-dotenv
openai>=1.0.0
google-generativeai>=0.8.0
geopy
Pillow
gunicorn
requests
